moved database connection to /server/db/index
