+++ server/api/index.js
@@ -1,0 +1,10 @@
+const router = require('express').Router()
+module.exports = router
+
+router.use('/game', require('./game'))
+
+router.use((req, res, next) => {
+  const error = new Error('Not Found')
+  error.status = 404
+  next(error)
+})
