+++ server/index.js
@@ -8,7 +8,7 @@ const PORT = process.env.PORT || 8080
 const app = express()
 const socketio = require('socket.io')
 
-const admin = require('firebase-admin')
+// const admin = require('firebase-admin')
 
 // const serviceAccount = require('path/to/serviceAccountKey.json')
 
@@ -24,18 +24,18 @@ module.exports = app
  */
 if (process.env.NODE_ENV !== 'production') require('../secrets')
 
-// Initialize database with global variables
-admin.initializeApp({
-  credential: admin.credential.cert({
-    projectId: process.env.project_id,
-    clientEmail: process.env.client_email,
-    privateKey: process.env.private_key.replace(/\\n/g, '\n')
-  }),
-  databaseURL: 'https://guava-stackbox.firebaseio.com'
-})
+// // Initialize database with global variables
+// admin.initializeApp({
+//   credential: admin.credential.cert({
+//     projectId: process.env.project_id,
+//     clientEmail: process.env.client_email,
+//     privateKey: process.env.private_key.replace(/\\n/g, '\n')
+//   }),
+//   databaseURL: 'https://guava-stackbox.firebaseio.com'
+// })
 
-// Back-end firebase is admin.database()
-const database = admin.database()
+// // Back-end firebase is admin.database()
+// const database = admin.database()
 
 // Write from the back end
 
@@ -77,6 +77,7 @@ const createApp = () => {
   app.use(passport.initialize())
   app.use(passport.session())
 
+  app.use('/api', require('./api'))
   // static file-serving middleware
   app.use(express.static(path.join(__dirname, '..', 'public')))
